Add files via upload
BookPals is a book-themed social where the main entities are Users, Books, Authors and, obviously me, the admin. Here you should be able to follow the authors you like and get a notification whenever a new book is loaded in the database by the admin or the authors themselves. Users could choose to become "BookPals" and share their readings and their opinions about a book.
In this branch, BookPals is developed without the aid of frameworks (only HTML, CSS, JS and BOOTSTRAP are used) and there aren't any realistic data to see, only the view is showed. If you wanna see more realistic features, switch to other branches with frameworks or another repository where both front end and back end are developed.